Make the notification channel optional
My application doesn't need it, and probably a bunch more as well,
so this changes the interface so that you can just pass None as
the size, and the channel is never created or used.

This could probably be a bit better -- you could have a type that
supports notifications and a type that doesn't, avoiding the error
of asking to subscribe when there is no channel.